requirements.txt
absl-py==0.9.0
actionlib==1.12.0
altgraph==0.17
angles==1.9.11
audioread==2.1.8
auto-py-to-exe==2.7.3
beautifulsoup4==4.9.0
bleach==1.5.0
bondpy==1.8.3
bottle==0.12.18
bottle-websocket==0.2.9
camera-calibration==1.13.0
camera-calibration-parsers==1.11.13
catkin==0.7.20
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
controller-manager==0.15.1
controller-manager-msgs==0.15.1
cursor==1.3.4
cv-bridge==1.13.0
cycler==0.10.0
decorator==4.4.2
deepspeech==0.7.1
deepspeech-tflite==0.7.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
dlib==19.19.0
dynamic-reconfigure==1.6.0
Eel==0.11.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
future==0.18.2
gazebo-plugins==2.8.6
gazebo-ros==2.8.6
gencpp==0.6.2
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.12
gennodejs==2.0.1
genpy==0.6.9
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
halo==0.0.29
html5lib==0.9999999
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.9
image-geometry==1.13.0
imutils==0.5.3
interactive-markers==1.11.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kdl-parser-py==1.13.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
laser-geometry==1.6.4
librosa==0.7.2
llvmlite==0.32.0
log-symbols==0.0.14
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
message-filters==1.14.3
mpyg321==0.0.2
noisereduce==1.1.0
numba==0.49.0
numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
playsound==1.2.2
protobuf==3.11.3
PyAudio==0.2.11
pycparser==2.20
pygame==1.9.6
PyInstaller==3.6
PyJWT==1.7.1
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyqtgraph==0.10.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-engineio==3.12.1
python-qt-binding==0.3.6
python-socketio==4.5.1
python-vlc==3.0.9113
pyttsx3==2.87
pytz==2019.3
PyYAML==5.3.1
qt-dotgraph==0.3.16
qt-gui==0.3.16
qt-gui-cpp==0.3.16
qt-gui-py-common==0.3.16
requests==2.23.0
resampy==0.2.2
resource-retriever==1.12.5
rosbag==1.14.3
rosboost-cfg==1.14.7
rosclean==1.14.7
roscreate==1.14.7
rosgraph==1.14.3
roslaunch==1.14.3
roslib==1.14.7
roslint==0.11.2
roslz4==1.14.3
rosmake==1.14.7
rosmaster==1.14.3
rosmsg==1.14.3
rosnode==1.14.3
rosparam==1.14.3
rospy==1.14.3
rosservice==1.14.3
rostest==1.14.3
rostopic==1.14.3
rosunit==1.14.7
roswtf==1.14.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.14
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.7
scikit-learn==0.22.2.post1
scipy==1.4.1
sensor-msgs==1.12.7
six==1.14.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
SoundCard==0.3.3
sounddevice==0.3.15
SoundFile==0.10.3.post1
soupsieve==2.0
spinners==0.0.24
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tf==1.12.0
tf-conversions==1.12.0
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
tkintertable==1.3.2
topic-tools==1.14.3
tqdm==4.45.0
urdfdom-py==0.4.1
urllib3==1.25.9
webrtcvad==2.0.10
websocket-client==0.48.0
Werkzeug==1.0.1
wget==3.2
whichcraft==0.6.1
xacro==1.13.5
